% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bat.diet.hierarchy}
\alias{bat.diet.hierarchy}
\title{Bat diet hierarchy}
\format{
A data frame with 40 rows and 2 columns.
}
\usage{
bat.diet.hierarchy
}
\description{
Hierarchy table indicating the relationship between samples and their
 respective parent groups.
}
\keyword{datasets}
