% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bat.diet.otutable}
\alias{bat.diet.otutable}
\title{Bat diet OTU table}
\format{
A data frame with 363 rows and 40 species.
}
\usage{
bat.diet.otutable
}
\description{
An OTU table containing the absolute read abundances of 363 OTUs in 40
 faecal samples from 8 different bat species.
}
\keyword{datasets}
