% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bat.diet.tree}
\alias{bat.diet.tree}
\title{Bat diet OTU tree}
\format{
A phylo object with 363 tips and 362 internal nodes.
}
\usage{
bat.diet.tree
}
\description{
Phylogenetic tree built from the representative sequences of the 363 OTUs
 included in the 'bat.diet.otutable' data set.
}
\keyword{datasets}
