\name{copy.filt}
\alias{copy.filt}
\title{Copy number threshold filtering}
\usage{beta.dis(otutabe,threshold)}
\arguments{
  \item{otutable}{A matrix indicating the relative abundances of multiple samples. Columns should be samples and rows OTUs.}
  \item{threshold}{}
}
\details{}
\examples{}
\references{}
