% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_nmds.R
\name{dis_nmds}
\alias{dis_nmds}
\title{Dissimilarity NMDS plot}
\usage{
dis_nmds(distance, hierarchy, colour, plot, centroids, labels, runs)
}
\arguments{
\item{distance}{Matrix of pairwise dissimilarities, usually one of the matrices listed in the output object of the pair_dis() function.}

\item{hierarchy}{The first column containing the sample names while the second containing the groups names. If provided, dots are coloured according to groups and group centroids can also be visualised.}

\item{colour}{The number of vector items (colours, e.g. '#34k235'), must be of length one if no hierarchy table is added, or must equal the number of groups if the hierarchy table is provided.}

\item{plot}{Whether to plot a NMDS or a Shepard plot. Default: "NMDS".}

\item{centroids}{Whether to link sample dots with group centroids or not. A hierarchy table is necessary to draw centroids. Default: FALSE}

\item{labels}{Whether to print sample or group labels or both. A hierarchy table is necessary to plot grpup names. Default: "none".}

\item{runs}{Number of iterations for the NMDS function. Default: 100.}
}
\value{
An NMDS or Shepard plot.
}
\description{
Visualisation of pairwise dissimilarities
}
\details{
Dissimilarity NMDS plot
}
\examples{
data(bat.diet.otutable)
data(bat.diet.tree)
data(bat.diet.hierarchy)
pairdisres <- pair_dis(bat.diet.otutable,qvalue=0,hierarchy=bat.diet.hierarchy)
dis_nmds(pairdisres$L1_CqN)
dis_nmds(pairdisres$L1_CqN,hierarchy=bat.diet.hierarchy, centroids=TRUE)
dis_nmds(pairdisres$L1_CqN,hierarchy=bat.diet.hierarchy, centroids=TRUE, labels="group")
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources, 19, 804-817.\cr\cr
Chao, A., Chiu, C.H., & Hsieh, T. C. (2012). Proposing a resolution to debates on diversity partitioning. Ecology, 93, 2037-2051.\cr\cr
Jost, L. (2007). Partitioning diversity into independent alpha and beta components. Ecology, 88, 2427-2439.
}
\seealso{
\code{\link{pair_dis}}, \code{\link{beta_dis}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{diversity}
\keyword{hill}
\keyword{numbers}
\keyword{ordination}
\keyword{pairwise}
\keyword{partitioning}
\keyword{plot}
