% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.profile.R
\name{div.profile}
\alias{div.profile}
\title{Diversity profile}
\usage{
div.profile(abund, qvalues, tree, values, hierarchy, level, colour, log)
}
\arguments{
\item{abund}{A vector or a matrix/data.frame indicating the relative abundances of one or multiple samples, respectively. If a matrix/data.frame is provided, columns must refer to samples and rows to OTUs.}

\item{qvalues}{A vector of sequential orders of diversity (default from 0 to 5). order=seq(from = 0, to = 5, by = (0.1))}

\item{tree}{A tree of class 'phylo'. The tip labels must match the names of the vector values (if one sample) or matrix rows (if multiple samples).}

\item{values}{If 'TRUE' the Hill numbers for each order of diversity for each sample or group (if hierarchy table is provided) are printed on screen. values="FALSE".}

\item{hierarchy}{A two-column matrix indicating the relation between samples (first column) and groups (second column).}

\item{level}{Only if hierachy is used. If 'alpha', alpha diversity of the groups is plotted instead of gamma diversity. level="gamma"}

\item{colour}{The number of vector items (colours, e.g. '#34k235'), must equal the number of samples or groups that are intended to plot.}

\item{log}{If 'TRUE' the Hill numbers are transformed to the logarithmic scale. This is useful when there are large differences between q values (e.g. sharp drop from q=0 to q=1), which might complicate visualization. log="FALSE"}
}
\value{
A figure or a table of sequential diversity values.
}
\description{
Create diversity profiles of a single or multiple samples displayed independently or aggregated in groups.
}
\details{
Diversity profile
}
\examples{
div.profile(abund=otu.table[,1],qvalues=seq(from = 0, to = 5, by = (0.1)),tree=tree,values="TRUE")
div.profile(vector)
div.profile(otu.table,hierarchy=hierarchy.table)
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources. Early view.\cr\cr
Chao, A., Chiu, C.‐H., & Jost, L. (2014). Unifying species diversity, phylo‐ genetic diversity, functional diversity, and related similarity and dif‐ ferentiation measures through hill numbers. Annual Review of Ecology Evolution and Systematics, 45, 297–324.
}
\seealso{
\code{\link{hill.div}}, \code{\link{div.part}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{alpha}
\keyword{beta}
\keyword{gamma}
\keyword{hill}
