% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.test.plot.R
\name{div.test.plot}
\alias{div.test.plot}
\title{Diversity test plotting}
\usage{
div.test.plot(divtest, chart, colour, posthoc, threshold)
}
\arguments{
\item{divtest}{Object outputed by the div.test() function}

\item{chart}{Chart type, either 'box' for boxplot, 'jitter' for jitter plot or 'violin' for violin plot. chart="box"}

\item{colour}{The number of vector items (colours, e.g. '#34k235'), must equal the number of groups that are intended to plot.}

\item{posthoc}{If 'TRUE' pairwise p-values of the posthoc analyses will be ploted. It requires the div.test() object to contain posthoc results.}

\item{threshold}{Maximum p-value to show in pairwise posthoc results (usually 0.05, but could be any other number between 0 an 1). P-values above the threshold will not be showed.}
}
\value{
Chart of (mean) diversities of contrasting groups with optional posthoc results.
}
\description{
Plot of diversity comparison between groups of samples
}
\details{
Diversity test plotting
}
\examples{
contrast.div.q0 <- div.test(otu.table,qvalue=0,hierarchy=hierarchy.table)
div.test.plot(contrast.div.q0,chart="jitter")
div.test.plot(contrast.div.q0,chart="violin")
div.test.plot(div.test.result,chart="jitter",posthoc=TRUE,threshold=0.5)
}
\seealso{
\code{\link{div.test}}, \code{\link{hill.div}}, \code{\link{div.part}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{chart}
\keyword{comparison}
\keyword{hill}
\keyword{numbers}
