% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.data.R
\name{match.data}
\alias{match.data}
\title{Match data}
\usage{
match.data(otutable, tree, output)
}
\arguments{
\item{otutable}{A vector or a matrix/data.frame indicating the relative abundances of one or multiple samples, respectively. If a matrix/data.frame is provided, columns must refer to samples and rows to OTUs.}

\item{tree}{An ultrametic tree of class 'phylo'.}

\item{output}{Whether to output a filtered OTU table (matrix) or a filtered OTU tree (phylo).}
}
\description{
Filter count tables and OTU/ASV phylogenetic trees to match OTUs/ASVs present in both data files..
}
\details{
Match data
}
\examples{
match.data(bat.diet.otutable,bat.diet.tree,output="otutable")
match.data(bat.diet.otutable,bat.diet.tree,output="tree")
}
\seealso{
\code{\link{hill.div}}, \code{\link{index.div}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{ASV}
\keyword{OTU}
\keyword{names}
\keyword{tree}
