% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair.dis.plot.R
\name{pair.dis.plot}
\alias{pair.dis.plot}
\title{Pairwise dissimilarity plot}
\usage{
pair.dis.plot(distance, hierarchy, type, level, colour, magnify)
}
\arguments{
\item{distance}{Matrix of pairwise dissimilarities, usually one of the matrices listed in the output object of the pair.dis() function.}

\item{hierarchy}{The first column lists the sample names while the second lists the groups. If provided, group profiles are plotted instead of individual profiles.}

\item{type}{Whether to plot a NMDS or qgraph chart. type="NMDS".}

\item{level}{If '1', samples are not grouped (each sample is a different colour), while if '2', samples are coloured by groups. level=1.}

\item{colour}{he number of vector items (colours, e.g. '#34k235'), must equal the number of samples or groups that are intended to plot.}

\item{magnify}{Only relevant for qgraph. Whether the pairwise dissimilarity values are transformed to 0-1 scale, 0 corresponding to the minimum dissimilarity and 1 to the maximum dissimilarity value. magnify=FALSE.}
}
\value{
An NMDS or network plot.
}
\description{
Visualisation of pairwise dissimilarities
}
\details{
Pairwise dissimilarity plot
}
\examples{
pair.div.q0.L2 <- pair.dis(otu.table,qvalue=0,hierarchy=hierarchy.table,level="2")
pair.dis.plot(pair.div.q0.L2$L2_CqN,hierarchy=hierarchy.table,type="NMDS",level=2)
pair.dis.plot(pair.div.q0.L2$L2_CqN,hierarchy=hierarchy.table,type="qgraph",level=2,magnify=TRUE)
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources, 19, 804-817.
Chao, A., Chiu, C.‐H., & Hsieh, T. C. (2012). Proposing a resolution to de‐ bates on diversity partitioning. Ecology, 93, 2037–2051
Jost, L. (2007). Partitioning diversity into independent alpha and beta components. Ecology, 88, 2427–2439.
}
\seealso{
\code{\link{pair.dis}}, \code{\link{beta.dis}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{diversity}
\keyword{hill}
\keyword{numbers}
\keyword{pairwise}
\keyword{partitioning}
