\name{phylo.to.phylog}
\alias{phylo.to.phylog}
\title{Transform tree from phylo to phylog format}
\usage{phylo.to.phylog(tree)}
\arguments{
  \item{tree}{A tree of class 'phylo'.}
}
\details{A tree of class 'phylog' (ade4) is necessary to run iNextPD. This function transforms a 'phylo' object into a 'phylog' object.}
