\name{beta.dis}
\alias{beta.dis}
\title{(Dis)similarity measures derived from Hill number beta diversities}
\usage{}
\arguments{
  \item{beta}{A numeric beta value or a 'div.part' function output object.}
  \item{qvalue}{A positive number (>0). It can be an integer or contain decimals. The argument is ovewritten if a 'div.part' object is used.}
  \item{N}{Integer. Sample size, the number of sampling units to be used to compute the (dis)similarity measures. The argument is ovewritten if a 'div.part' object is used}.
  \item{metric}{A character object containing "C", "U", "V" or "S". See the explanation of each metric below. metric="C".}
  \item{type}{A character object containing either "similarity" or "dissimilarity". Defines the 'orientation' of the metric. type="dissimilarity"}
}
