% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.part.R
\name{div.part}
\alias{div.part}
\title{Diversity partitioning}
\usage{
div.part(otutable, qvalue, tree, type, hierarchy)
}
\arguments{
\item{otutable}{An OTU table (matrix/data.frame) indicating the absolute or relative OTU abundances of multiple samples. Columns must refer to samples and rows to OTUs.}

\item{qvalue}{A positive number, usually between 0 and 5, but most commonly 0, 1 or 2. It can be an integer or contain decimals.}

\item{tree}{A phylogenetic tree of class 'phylo'. The tip labels must match the row names in the OTU table. Use the function match.data() if the OTU names do not match.}

\item{type}{Either "abundance" for abundance-based Hill number computation or "incidence" for incidence-based Hill number computation. Incidence-based Hill number computation requires a hierarchy table.}

\item{hierarchy}{A two-column matrix indicating the relation between samples (first column) and groups (second column).}
}
\value{
A list object containing general information about the diversity partitioning and L1 (level 1 = alpha), L2 (level 2 = gamma) and beta diversities.
}
\description{
Diversity partitioning based on Hill numbers. the Hill numbers framework enables the diversity of a system to be partitioned following the multiplicative definition (Jost, 2007; Chao, Chiu, & Hsieh, 2012). Alpha diversity is obtained by computing the Hill numbers from the averaged basic sums of the samples, while gamma diversity is obtained by taking the average of OTU relative abundances across samples, and then computing the Hill numbers of the pooled system. The division of gamma diversity by alpha diversity yields the beta diversity, which quantifies how many times richer an entire system is in effective OTUs (gamma diversity) than its constituent samples are on average (alpha diversity). However, the Hill numbers beta diversity can also be considered an actual diversity value, as the same metric also measures the effective number of equally large and completely distinct samples in a system.
}
\details{
Diversity partitioning (based on Hill numbers)
}
\examples{
div.part(otu.table,qvalue=1)
div.part(otu.table,qvalue=1,type="abundance")
div.part(otu.table,qvalue=0,tree=tree)
div.part(otu.table,qvalue=0,type="incidence",hierarchy=hierarchy.table)
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources, 19, 804-817.
Chao, A., Chiu, C.‐H., & Hsieh, T. C. (2012). Proposing a resolution to de‐ bates on diversity partitioning. Ecology, 93, 2037–2051
Jost, L. (2007). Partitioning diversity into independent alpha and beta components. Ecology, 88, 2427–2439.
}
\seealso{
\code{\link{div.part}}, \code{\link{gamma.div}}, \code{\link{match.data}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{alpha}
\keyword{beta}
\keyword{gamma}
\keyword{hill}
