% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.profile.plot.R
\name{div.profile.plot}
\alias{div.profile.plot}
\title{Diversity profile plot}
\usage{
div.profile.plot(profile, colour, log, legend)
}
\arguments{
\item{profile}{A div.profile() object or a vector/matrix containg diversity profile(s), with columns indicating samples/groups and rows indicating orders of diversity (q-values).}

\item{colour}{The number of vector items (colours, e.g. '#34k235'), must equal the number of samples or groups that are intended to plot.}

\item{log}{If 'TRUE' the Hill numbers are transformed to the logarithmic scale. This is useful when there are large differences between q values (e.g. sharp drop from q=0 to q=1), which might complicate visualization. log="FALSE"}

\item{legend}{Whether to display the legend (TRUE) or not (FALSE) in diversity profiles containing multiple samples/groups. Default TRUE in multi-sample charts.}
}
\value{
A diversity profile plot.
}
\description{
Plot diversity profiles of a single or multiple samples displayed independently or aggregated in groups.
}
\details{
Diversity profile plot
}
\examples{
div.profile.plot(profile)
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources. Early view.\cr\cr
Chao, A., Chiu, C.‐H., & Jost, L. (2014). Unifying species diversity, phylo‐ genetic diversity, functional diversity, and related similarity and dif‐ ferentiation measures through hill numbers. Annual Review of Ecology Evolution and Systematics, 45, 297–324.
}
\seealso{
\code{\link{div.profile}}, \code{\link{hill.div}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{alpha}
\keyword{beta}
\keyword{gamma}
\keyword{hill}
