% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.profile.R
\name{div.profile}
\alias{div.profile}
\title{Diversity profile}
\usage{
div.profile(abund, qvalues, tree, hierarchy, level)
}
\arguments{
\item{abund}{A vector or a matrix/data.frame indicating the relative abundances of one or multiple samples, respectively. If a matrix/data.frame is provided, columns must refer to samples and rows to OTUs.}

\item{qvalues}{A vector of sequential orders of diversity (default from 0 to 5). order=seq(from = 0, to = 5, by = (0.1))}

\item{tree}{A tree of class 'phylo'. The tip labels must match the names of the vector values (if one sample) or matrix rows (if multiple samples).}

\item{hierarchy}{A two-column matrix indicating the relation between samples (first column) and groups (second column).}

\item{level}{Whether to compute alpha or gamma diversities of the system or the groups specified in the hierarchy table.}
}
\value{
A vector or matrix containing diversity values at different orders of diversity (as specified in qvalues).
}
\description{
Create diversity profiles of a single or multiple samples displayed independently or aggregated in groups.
}
\details{
Diversity profile
}
\examples{
div.profile(abund=otu.table[,1],qvalues=seq(from = 0, to = 5, by = (0.1)),tree=tree)
div.profile(vector)
div.profile(otu.table,hierarchy=hierarchy.table,level="gamma")
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources. Early view.\cr\cr
Chao, A., Chiu, C.‐H., & Jost, L. (2014). Unifying species diversity, phylo‐ genetic diversity, functional diversity, and related similarity and dif‐ ferentiation measures through hill numbers. Annual Review of Ecology Evolution and Systematics, 45, 297–324.
}
\seealso{
\code{\link{div.profile.plot}}, \code{\link{hill.div}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{alpha}
\keyword{beta}
\keyword{gamma}
\keyword{hill}
