% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.test.plot.R
\name{div.test.plot}
\alias{div.test.plot}
\title{Diversity test plotting}
\usage{
div.test.plot(divtest, chart, colour, stat, comb, symbol, flip)
}
\arguments{
\item{divtest}{Object outputed by the div.test() function}

\item{chart}{Chart type, either 'box' for boxplot, 'jitter' for jitter plot or 'violin' for violin plot. chart="box"}

\item{colour}{The number of vector items (colours, e.g. '#34k235'), must equal the number of groups that are intended to plot.}

\item{stat}{If 'TRUE' pairwise mean comparison significance values will be ploted.}

\item{comb}{List of pairwise combinations. If missing all combinations will be plotted. e.g. list(c("Myotis myotis","Myotis capaccinii"),c("Myotis myotis","Myotis daubentonii")).}

\item{symbol}{If 'TRUE' symbols rather than p-values will be shown. ns: p > 0.05; *: p <= 0.05; **: p <= 0.01; ***: p <= 0.001; ****: p <= 0.0001.}

\item{flip}{If 'TRUE' the chart will be flipped 90 degrees.}
}
\value{
Chart of (mean) diversities of contrasting groups.
}
\description{
Plot of diversity comparison between groups of samples
}
\details{
Diversity test plotting
}
\examples{
contrast.div.q0 <- div.test(otu.table,qvalue=0,hierarchy=hierarchy.table)
div.test.plot(contrast.div.q0,chart="jitter")
div.test.plot(contrast.div.q0,chart="violin")
div.test.plot(div.test.result,chart="jitter",stat=TRUE,flip=TRUE)
div.test.plot(div.test.result,stat=TRUE,comb=list(c("Myotis myotis","Myotis capaccinii")),symbol=TRUE)
}
\seealso{
\code{\link{div.test}}, \code{\link{hill.div}}, \code{\link{div.part}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{chart}
\keyword{comparison}
\keyword{hill}
\keyword{numbers}
