% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.test.R
\name{div.test}
\alias{div.test}
\title{Diversity test}
\usage{
div.test(otutable, qvalue, hierarchy, tree, posthoc)
}
\arguments{
\item{otutable}{A matrix indicating the relative abundances of multiple samples. Columns should be samples and rows OTUs.}

\item{qvalue}{A positive integer or decimal number (>=0), usually between 0 and 3.}

\item{hierarchy}{A two-column matrix indicating the relation between samples (first column) and groups (second column).}

\item{tree}{A phylogenetic tree of class 'phylo'. The tip labels must match the row names in the OTU table. Use the function match.data() if the OTU names do not match.}

\item{posthoc}{Whether to run post hoc pairwise analyses or not. If TRUE, an ANOVA will be complemented with a Tukey test and a Kruskal-Wallis test will be complemented with a Dunn test.}
}
\value{
A statistical test output.
}
\description{
Diversity comparison test between groups of samples
}
\details{
Diversity test
}
\examples{
div.test(otu.table,qvalue=0,hierarchy=hierarchy.table)
div.test(otu.table,qvalue=1,hierarchy=hierarchy.table,tree=tree)
div.test(otu.table,2,hierarchy.table,tree)
div.test(otu.table,qvalue=1,hierarchy=hierarchy.table,posthoc=TRUE)
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources. Early view.\cr\cr
Chao, A., Chiu, C.‐H., & Jost, L. (2014). Unifying species diversity, phylo‐ genetic diversity, functional diversity, and related similarity and dif‐ ferentiation measures through hill numbers. Annual Review of Ecology Evolution and Systematics, 45, 297–324.
}
\seealso{
\code{\link{hill.div}}, \code{\link{div.part}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{comparison}
\keyword{hill}
\keyword{numbers}
