% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill.div.R
\name{hill.div}
\alias{hill.div}
\title{Hill numbers computation}
\usage{
hill.div(abund, qvalue, tree, type)
}
\arguments{
\item{abund}{A vector or a matrix/data.frame indicating the relative abundances of one or multiple samples, respectively. If a matrix/data.frame is provided, columns must refer to samples and rows to OTUs.}

\item{qvalue}{A positive integer or decimal number (>=0), usually between 0 and 3.}

\item{tree}{An ultrametic tree of class 'phylo'. The tip labels must match the names of the vector values (if one sample) or matrix rows (if multiple samples).}

\item{type}{Either "abundance" for abundance-based Hill number computation or "incidence" for incidence-based Hill number computation. In the latter case, the 'abund' object needs to be an OTU table, as incidence-based Hill number computation requires multiple samples. Default: type="abundance".}
}
\description{
Compute neutral or phylogenetic Hill numbers from a single sample (vector) or count table (matrix). Hill numbers or numbers equivalents of diversity indices are diversity measures that compute diversity in effective number of OTUs, i.e. the number of equally abundant OTUs that would be needed to give the same value of diversity.
}
\details{
Hill numbers computation
}
\examples{
hill.div(otu.vector,0)
hill.div(otu.table,1)
hill.div(otu.table,1,tree)
hill.div(otu.table,1,tree,type="incidence")
hill.div(otu.table,qvalue=2,tree=tree)
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources. Early view.\cr\cr
Jost, L. (2006). Entropy and diversity. Oikos, 113, 363–375.\cr\cr
Hill, M. O. (1973). Diversity and evenness: a unifying notation and its con‐ sequences. Ecology, 54, 427–432.
}
\seealso{
\code{\link{index.div}}, \code{\link{div.part}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{ASV}
\keyword{OTU}
\keyword{diversity}
