% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.depth.R
\name{tree.depth}
\alias{tree.depth}
\title{Tree depth}
\usage{
tree.depth(tree, abund)
}
\arguments{
\item{tree}{A phylogenetic tree of class 'phylo'. The tip labels must match the row names in the OTU table. Use the function match.data() if the OTU names do not match.}

\item{abund}{A vector or a matrix/data.frame indicating the relative abundances of one or multiple samples, respectively. If a matrix/data.frame is provided, columns must refer to samples and rows to OTUs.}
}
\value{
A tree depth value
}
\description{
Computes phylogenetic tree depth based from a phylogenetic tree and a vector of (relative) abundances.
}
\details{
Tree depth
}
\examples{
tree.depth(tree=otu.tree,abund=otu.table)
tree.depth(otu.tree,otu.table)
}
\seealso{
\code{\link{div.part}}, \code{\link{gamma.div}}, \code{\link{match.data}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{phylogeny}
\keyword{tree}
